requirements.txt
langchain
langchain_community
langchain-openai
langchain-groq
python-dotenv
langchain-google-genai
langchain-core
pypdf 
faiss-cpu 
sentence-transformers
langchain-text-splitters
langchain-huggingface